+++ packages/twenty-server/src/modules/connected-account/refresh-token-manager/google-api-refresh-access-token.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
-import { GoogleAPIRefreshAccessTokenService } from 'src/modules/connected-account/services/google-api-refresh-access-token/google-api-refresh-access-token.service';
+import { GoogleAPIRefreshAccessTokenService } from 'src/modules/connected-account/refresh-token-manager/google-api-refresh-access-token.service';
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
 import { MessagingCommonModule } from 'src/modules/messaging/common/messaging-common.module';
